Jana/src-tauri/resources/server/requirements.txt
Flask
flask-cors
openai-whisper
requests
kokoro-onnx
soundfile